+++ CHANGELOG.md
@@ -10,6 +10,9 @@ about writing changes to this log.
 
 ## [Unreleased]
 
+- [PR-353](https://github.com/itk-dev/naevnssekretariatet/pull/353)
+  Adds `referenceNumber` to `CasePartyRelation`
+  and usage of this during hearing
 - [PR-352](https://github.com/itk-dev/naevnssekretariatet/pull/352)
   Fixes digital post creation for inspection letters
 - [PR-350](https://github.com/itk-dev/naevnssekretariatet/pull/350)